encapsulate transaction-history event
